+++ sopa/io/__init__.py
@@ -2,6 +2,7 @@ from .imaging import macsima, phenocycler, hyperion, ome_tif
 from .explorer import write
 from .standardize import write_standardized
 from .transcriptomics import merscope, xenium, cosmx
+from .histopathology import wsi, wsi_autoscale
 from .report import write_report
 
 from ..utils.data import blobs, uniform