+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## [1.0.5] - tbd
+## [1.0.5] - 2024-03-01
 
 ### Changed
 - Faster image writing for the Xenium Explorer (about x5 speedup)
@@ -7,7 +7,7 @@
 
 ### Added
 - Support for python 3.9 to 3.12 (we still recommend `python==3.10`)
-- Support WSI analysis (reader, tissue segmentation, patch embedding)
+- Support WSI analysis: reader, tissue segmentation, patch embedding (tutorials coming soon)
 - Supporting multiple region-of-interest queries
 - Can load a custom cellpose model using the `pretrained_model`/`model_type` argument